requirements.txt
boto3
duckdb
polars
pandas
numpy
yfinance
pyarrow
plotly
pyyaml
pytest
rich
streamlit
requests
vaderSentiment
scikit-learn
scipy
python-dotenv
